Home details page
Breaking out the Button component to be able to support being passed in a boolean prop "showText". This will allow a button to show a text and image with flex space-between or just an image in example to the Resume button. Still need to build out the "just image" button to use for the resume button.

Ability to download resume

rearranging home details

Download button styled